+++ anyblok/model.py
@@ -243,6 +243,24 @@ class Model:
                 ev.append(val)
 
     @classmethod
+    def apply_sqlalchemy_event_listner(cls, attr, method, registry, namespace,
+                                       base, properties):
+        """declare in the registry the sqlalchemy event
+
+        :param attr: name of the attibute
+        :param method: method pointer
+        :param registry: the current  registry
+        :param namespace: the namespace of the model
+        :param base: One of the base of the model
+        :param properties: the properties of the model
+        """
+        if not hasattr(method, 'is_an_sqlalchemy_event_listener'):
+            return
+        elif method.is_an_sqlalchemy_event_listener is True:
+            registry._sqlalchemy_known_events.append(
+                (method.sqlalchemy_listener,  ModelAttribute(namespace, attr)))
+
+    @classmethod
     def detect_hybrid_method(cls, attr, method, registry, namespace, base,
                              properties):
         """ Find the sqlalchemy hybrid methods in the base to save the
@@ -311,6 +329,8 @@ class Model:
             new_type_properties.update(apply_cache(
                 attr, method, registry, namespace, base, properties))
             cls.apply_event_listner(
+                attr, method, registry, namespace, base, properties)
+            cls.apply_sqlalchemy_event_listner(
                 attr, method, registry, namespace, base, properties)
             cls.detect_hybrid_method(
                 attr, method, registry, namespace, base, properties)